+++ src/main/resources/db/migration/V1_1__players_table.sql
@@ -1,8 +1,8 @@
 -- Players
 CREATE TABLE players (
   id       BIGSERIAL PRIMARY KEY,
-  username VARCHAR NOT NULL UNIQUE,
-  password VARCHAR NOT NULL,
-  rating   INT     NOT NULL,
-  active   BOOLEAN NOT NULL DEFAULT FALSE
+  username VARCHAR          NOT NULL UNIQUE,
+  password VARCHAR          NOT NULL,
+  rating   DOUBLE PRECISION NOT NULL,
+  active   BOOLEAN          NOT NULL DEFAULT FALSE
 );
